Fixed conversion of struct timeval to seconds   - to_sec() performs conversion to double type   - now 500 microseconds will be represented as 0.0005 and must be rounded to 0.001   - added unit-test with check of conversion to string with rounding to lower precision

+++ analyzers/src/breakdown/latencies.cpp
@@ -87,8 +87,8 @@ void Latencies::set_range(const timeval& t)
     }
 }
 
-long double NST::breakdown::to_sec(const timeval& val)
+double NST::breakdown::to_sec(const timeval& val)
 {
-    return static_cast<long double>(val.tv_sec) + val.tv_usec / 1000000.0L;
+    return static_cast<double>(val.tv_sec) + static_cast<double>(val.tv_usec) / 1000000.0;
 }
 //------------------------------------------------------------------------------
